Reduce ParseProjectJob's memory usage
d->filesToParse is never used after all its elements are queued for
parsing. If parseAllProjectSources=true is passed to ParseProjectJob(),
this set contains almost as any elements as the number of files in the
project. If at least one of the project's documents is open, this set is
already detached, so setting it to {} frees memory. Otherwise, setting
to {} prevents detaching in case the project's file set changes later.