Sync CMake build type between cache and kdevelop settings
We store the cmake build type also in the kdevelo project settings
in order to recreate the build dir on demand if it gets wiped.
But when the user now manually changes it through the cmake cache
value for an existing build dir, the values could get out of sync.
Prevent that by keeping the two values in sync, i.e. once in the
list of cache values and once in the project setting drop down.

BUG: 429605