Keep include paths to build folders as project includes
Sadly, the order in which system and project includes are handled
is not as flexible as I hoped - system includes will always be handled
after the non-system includes.

Then, we try to auto-deduce system includes instead of relying on the
project manager for that distinction, which may lead to changed order
of include paths. In the kdevelop code base e.g. we often will show
parse errors for the debug categories, because the wrong `debug.h`
file is seen. Now this is fixed as the build dir includes are not
converted into system includes anymore and thus the order is correct.