Reactivate DBus connection to DrKonqi
Use QDBusServiceWatcher to watch for instances of DrKonqi.
KDevelop is then proposed as possible debugger of the crashed
application. This fixes the replacement of the deprecated
serviceOwnerChanged signal.
A unit test is added which simulates the expected behavior of
DrKonqi and checks if the PID of a crashed process is
correctly passed to KDevelop.