CMake: move KConfig keys to a header and reuse them in tests
Change the type of the global constants from QString to const char*
whenever possible to initialize them statically rather than dynamically.
In the few cases where const char* is unfeasible, replace the constants
with functions that return QString or QByteArray. Thus the possibility
of static initialization order fiasco and other issues with global
variables are eliminated.

These constants are used only as KConfigCore keys. Overloads of member
functions of KConfigCore classes, which take const QString& key, convert
the QString argument to const char*. So this change is beneficial to
performance.

Duplication in every translation unit of the UTF-8 and UTF-16 string
contents of a QStringLiteral and a QByteArrayLiteral in the new header
shouldn't be a problem, because this header is included only in one
source file and a test helper header, which in turn is included only in
test source files. So I haven't created cmakeconfiggroupkeys.cpp just to
de-inline the two *Literal-containing functions.

Remove the unused Config::Old::oldcmakeExecutableKey variable.

GIT_SILENT