OutputView: don't open a nonexistent file on click
Steps to reproduce the bug fixed by this commit:
1. Open kdevelop project in KDevelop.
2. Configure an in-tree build directory (e.g. kdevelop/build) with Ninja
   CMake generator. When GNU Make CMake generator or an out-of-tree
   build is used, compiler error/warning messages contain absolute paths
   to files. Only relative paths trigger the bug.
3. Modify any CMakeLists.txt file in the kdevelop project, e.g. insert a
   non-trailing space. Don't save the modified file yet (when some of
   the CMakeLists.txt files are saved, KDevelop runs cmake and thus
   interferes with reproducing the bug).
4. Modify e.g. kdevplatform/project/projectutils.cpp in a way that will
   trigger a compiler warning.
5. Build the kdevelop project (save the modified files if asked). The
   output in Build tool view should contain a CMake warning:
"CMake Warning at plugins/welcomepage/CMakeLists.txt:33 (message):"
   If not (i.e. because this warning no longer exists), insert a line
   like `message(WARNING "sample warning")` into a CMakeLists.txt in the
   3rd step. The compiler warning in projectutils.cpp should appear
   below, for example:
../kdevplatform/project/projectutils.cpp:30:6: warning: unused variable ‘a’ [-Wunused-variable]
6. Click on the projectutils.cpp warning line.

Observed result: a new empty file
kdevelop/../kdevplatform/project/projectutils.cpp is opened.
Expected result: the existing file
kdevelop/kdevplatform/project/projectutils.cpp is opened.

When CompilerFilterStrategyPrivate::pathForFile(
"../kdevplatform/project/projectutils.cpp") is called during the steps
to reproduce, m_buildDir equals "/path/to/kdevelop/build" and
m_currentDirs contains a single element "/path/to/kdevelop". This single
element is appended to m_currentDirs at the line
`d->putDirAtEnd( d->m_buildDir.parent() );` in
CompilerFilterStrategy::errorInLine(). Since m_currentDirs is not empty,
CompilerFilterStrategyPrivate::pathForFile() looks for the file in paths
relative to these directories. If the file does not exist in any of the
checked locations, this function returns the path relative to the first
directory in m_currentDirs. At this commit
CompilerFilterStrategyPrivate::pathForFile() returns the path relative
to m_buildDir as a fallback instead, which is the correct path for the
steps to reproduce above. Even if the path is not correct in some other
case, it shouldn't be worse than what was returned before this commit.

CCBUG: 381506