CompilerFilterStrategy: separate the CMake dir from compiler dirs
This commit fixes the same issue as the previous one - differently. The
fix in this commit is more complete: at the previous commit, if for some
reason the compiler-reported path also existed relative to the CMake
root directory, it was opened, even though it was not the right file.

In this commit we remember that the CMake root directory was appended to
m_currentDirs, then remove the directory from the container when
errors/warnings reported by CMake end and those reported by the compiler
begin. This way the (low) risk of opening the wrong file is eliminated
and performance is improved (because pathForFile() no longer looks for
each compiler-reported file in the wrong location).