Fix disappearing closed working set buttons
Fix the current behavior that working set buttons of
working sets that are active in another area (i.e. code,
review, debug) are not part of the current closed working
set collection.
New working sets are persistent by default, except for
newly created working sets for patch reviews.
When switching from the patch review to the code area, keep
the current working set open, if it is persistent (so if it is
not the dynamically created one).

BUG: 375446