CompilerFilterStrategy: remove seemingly obsolete code
On my GNU/Linux system, with GNU Make as the CMake generator, the
compiler always reports absolute paths in warnings and errors. With
Ninja generator, the paths are relative only when an in-tree build
directory is used, and they are always relative to the top-level build
directory. Though the "\\[.+%\\] Building .* object (.*)" regular
expression does not match Ninja output anyway (the corresponding lines
start with e.g. "[25/70]" rather than e.g. "[ 35%]").