Only show file name after reviewing a patch (revert d7dcc54d)
After switching the area back from a local patch review, only show the
file name, not the prettyFileName including project and path, i.e.
"myproject:src/main.cpp".

Commit d7dcc54d introduced a renaming of the title of views for files
in the code area that were recently used to review a patch. The review
plugin is triggered by using the "Review Patch" action in the context
menu of a text file's view. After switching back to the code area with
"Finish Review" the title of the file's view has changed.