Quick Open: sort and merge project files
Let N be m_projectFiles.size() at the beginning of
ProjectFileDataProvider::projectOpened(), that is, the number of files
that belong to projects that had been opened earlier. Let M be the
number of files in the just opened project (the number of times
KDevelop::forEachFile() invokes its callback, that is, including
duplicates). Currently ProjectFileDataProvider::projectOpened() inserts
project files one by one at the correct positions in a sorted QVector.
The worst-case complexity of this algorithm is O(M*(N+M)). Usually the
current algorithm works much faster - close to O(M*log(N+M)) - because
almost all insertions happen at the end or very close to the end of the
sorted QVector (as the project tree iterated over by
KDevelop::forEachFile() is close to sorted). But this much improved
performance holds only when there is a single open project or when
projects are opened in the order of their project directory paths in the
file system (so that almost all insertions still happen at the end or
very close to the end of m_projectFiles).

Reasonably assuming that additional memory is available, the worst-case
complexity of this commit's reimplementation of
ProjectFileDataProvider::projectOpened() is the sum of:
1) collecting the opened project's files in O(M);
2) std::stable_sort's O(M*log(M)) (with additional memory available);
3) std::inplace_merge's O(N+M);
4) removing duplicates in O(N+M).
The resulting worst-case complexity is O(N+M*log(M)). This is a vast
improvement over the current quadratic time in the worst-case, which
happens when multiple projects are opened in a certain order.

Timing the opening of large real-life projects and synthetic
BenchQuickOpen benchmarks indicate that the reimplementation performs
better most of the time.

The following table compares the average time in milliseconds spent in
ProjectFileDataProvider::projectOpened() at the two implementations.

version\projects     k=>W     W=>k=>q    S=> ( L=>W    |    W=>L )
previous commit     22+3058 1229+27+91  139+12040+56453+15255+1270
this commit         20+ 916  875+31+78   95+  591+ 1121+ 1071+ 776

        Legend
k=>W            - Open KDevelop with a session that contains kdevelop
                  and WebKit projects.
W=>k=>q         - Open WebKit, then kdevelop, then qt-creator project.
S=>(L=>W|W=>L)  - Open KDevelop with a session that contains Source,
                  LayoutTests and WebKit projects; close LayoutTests and
                  WebKit; open WebKit, then LayoutTests project (here
                  Source and LayoutTests are artificial projects in the
                  corresponding WebKit project's subdirectories, used to
                  benchmark opening heavily overlapping projects).

ProjectFileDataProvider::projectOpened() is called for projects in the
same order as they are listed in the legend above. In the table's
X+Y+... notation, X stands for the average time spent in the first call
to projectOpened() (first opened project), Y - in the second call to
projectOpened(), and so on.

The numbers of all/unique files in each tested project:
kdevelop:         5834/  4404
WebKit:         329562/321222
qt-creator:      20597/ 13887
Source:          38169/ 38169
LayoutTests:    218560/218560

The following table compares performance of the two implementations in
the affected BenchQuickOpen benchmarks. The numbers denote milliseconds
per iteration.

version\benchmark   1(1'000)    1(10'000)   3(1'000)    3(10'000)
previous commit     0.80        13          ~200        248
this commit         0.71         7.8        ~200        231

        Legend
1(1'000)    - benchProjectFileFilter_addRemoveProject(1000)
1(10'000)   - benchProjectFileFilter_addRemoveProject(10000)
3(1'000)    - benchProjectFileFilter_addRemoveProjects(1000)
3(10'000)   - benchProjectFileFilter_addRemoveProjects(10000)

The results of the 3(1'000) benchmark fluctuate so wildly that the
performance difference between the two implementations is indiscernible.