Quick Open: QVector => std::vector of project files
The fast prepend complication of QList/QVector in Qt 6 is not useful to
ProjectFileDataProvider::m_projectFiles as the elements are almost
always added and removed at the end. But the complexity of the new
implementation could be detrimental to Quick Open performance.

QVector's implicit sharing is not useful to m_projectFiles either. The
unnecessary copy-on-write poses a risk of accidental detaching.

This change brings a small performance improvement even with Qt 5.

The following table compares performance at the two commits in the
affected BenchQuickOpen::benchProjectFileFilter_addRemoveProject()
benchmark (in milliseconds per iteration).

version\benchmark   addRemoveProject(1000)  addRemoveProject(10000)
previous commit     0.38                    3.1
this commit         0.36                    3.0

The results of benchProjectFileFilter_addRemoveProjects() benchmark
fluctuate so wildly that the performance difference between the two
commits is indiscernible.