Inline default Path constructor
Inlining the default constructor improves performance when a collection
is resized to a large number of default-constructed elements. This
happens when ProjectFileDataProvider::reset() is called for the first
time after a large project is opened. My benchmarking indicates that
this change speeds up the first call to ProjectFileDataProvider::reset()
by about 4%.

The documentation of this constructor already promises "an empty,
invalid Path", so making Path::m_data non-empty in the default
constructor would be an API break. Therefore a potential ABI break
because of the inline constructor shouldn't be of concern.