Remove Path::clear()
Assigning Path{} should be more efficient most of the time. The absence
of the clear() member function forces using the usually faster
alternative. QVector::clear() is very slow if the detaching happens,
because the elements are first copied, then destroyed. Assigning Path{}
is always extremely fast, because the default Path constructor is inline
and QVector's move assignment operator boils down to several raw pointer
assignments. The only situation where clear() could be faster is when
detaching does not happen and the QVector's capacity is reused. But this
scenario seems to be unlikely for Path.