Inline Path::operator<
The recent commit that introduced Path::compare() and reimplemented
Path::operator< in terms of this new member function has unexpectedly
slowed down TestPath::bench_hash(), which never even calls compare() or
operator<, by 10%. This slowdown can be triggered by adding any
non-inline member function with a simple enough implementation to Path.
Inlining the now one-line Path::operator< restores the benchmark's
performance to the previous level.

The benefit of the inlining is the self-documentation of the
comparison's case sensitivity. The downside is that the implementation
can no longer be optimized without an ABI break. This downside is not a
problem for KDevPlatform, because each minor release breaks the ABI, and
optimizations don't target patch releases.