Store active working set in session configuration
Instead of storing the active working set in the global
configuration (overriding the setting when loading a
different session), store it in the session config.

At startup only load the working set for the Code area.
When switching the area, a new working set is loaded
anyway, discarding the preloaded working set.