Cache failing lookups of compiler builtin defines/includes
When a broken compiler is configured, or we query a non-existing
language/standard combination, like `-std=gnu++17` for C or OpenCL,
then no defines or includes are obtained. In such a situation, we would
keep trying to get the data repeatedly instead of caching the -
empty - result of the failed attempt.