Set GIT_OPTIONAL_LOCKS=0 for all git subprocesses
This should prevent kdevelop from interferring with other git
processes that are run manually by the user in e.g. a separate
terminal. Most notably, it should fix the pesky errors of this kind:

    fatal: Unable to create '....git/index.lock': File exists.

Support for this environment variable was added upstream in
version 2.15.0 from what I can see in [1].

[1]: https://github.com/git/git/commit/27344d6a6c8056664966e11acf674e5da6dd7ee3

BUG: 379219