Better cache compiler builtin defines and includes
We used to have a cache based on the compiler arguments in place.
But often, the compiler arguments change, yet the resulting builtin
defines and includes are going to be the same. This scenario was
not properly taken into account and thus we ended up asking the
compiler multiple times for the same results. By improving the
cache hit ratio here, we can shave off another large chunk of
where the main thread is getting blocked when opening a large dataset.