Fix crash when storing working sets area split-view config
When storing working sets commit 89594aed introduced a crash, which happens
when the code tries to find the parent splitter while the view doesn't have
an initialized widget. In this case calling `view->widget()` method tries
to create a new widget, but fails with an assert in `EditorViewWatcher::viewCreated`
because no parent for the widget is given.

BUG: 441058