IndexedString: check shouldDoDUChainReferenceCounting() earlier
I have written IndexedString statistics gathering code and run KDevelop
for several minutes. Turns out that shouldDoDUChainReferenceCounting()
calls in ref() and deref() were skipped less than 7% of the time because
the string is empty or single-char rarely. More than 99% of
shouldDoDUChainReferenceCounting() calls returned false, and probably
DUChainReferenceCounting::instance().count equaled zero in the calling
thread most of the time. When DUChainReferenceCounting::count == 0,
shouldDoDUChainReferenceCounting() returns false right away, without
even setting up the loop, thanks to Q_UNLIKELY.

Based on the gathered usage statistics, it should be faster to call
shouldDoDUChainReferenceCounting() first and skip empty and single-char
checks 99% of the time.

Moving the shouldDoDUChainReferenceCounting() calls out of ref() and
deref() allows to get rid of repeated calls to this function in the copy
assignment operator and the swap() overload.

The gathered statistics indicates that the two string arguments of the
swap() overload are equal 96% of the time => check for m_index equality
at the beginning of the swap() definition.

Affected average benchmark results before and at this commit in
milliseconds per iteration:
(benchmark ::function "data row")               Before      At
I BenchHashes
    ::insert()
        "qhash-1000"                            0.16        0.17
        "unordered_map-10000"                   1.8         1.9
        "qhash-10000"                           1.8         1.9
II BenchIndexedString
    ::bench_length()                            5.5         5.6
    ::bench_qhashIndexedString()                0.68        0.65
    ::bench_qSet()                              2.4         2.5
    ::bench_destroy()                           64          66
III BenchQuickOpen
    ::benchProjectFile_swap()                   0.000024    0.000027
    ::benchProjectFileFilter_addRemoveProject()
        "1000"                                  0.49        0.46
        "10000"                                 4.1         4.0
    ::benchProjectFileFilter_files()
        "1000, 10000"                           182         190
        "10000, 10000"                          396         412
        "8 projects"                            270         282
        "8 projects and one open file"          273         285
IV TestIdentifier
    ::benchQidCopyPush()                        0.00074     0.00076

Some benchmarks became faster but many became slower. The reason for the
slowdown is that the reduced number of calls to ref() and deref()
prompted GCC to inline the two editRepo() template instantiations in
these functions. This is a reversal of the generated code deduplication
caused by the recent addition of IndexedString swap() overload. The
clear performance degradation of the inlining suggests that the code
should be modified in a way that discourages inlining of editRepo().
This deserves a separate commit.