Fix the app shell script to run with zsh
Do not repeat shell test in the kdevplatform_shell_environment.sh file,
but instead detect the shell which runs the script via ps and $$. The
testing and shell selection is done by the app/kdevelop! script.

Fix tests for executables: which is a builtin in zsh and output
redirestion does not work with it. Replace such tests with command -v,
for which POSIX defines behaviour and return values.

BUG: 442481
FIXED-IN: 5.7.0