TestCodeCompletion::testCompletionPriority: QVERIFY => QCOMPARE
With this change the test's output tells which of the two equalities
doesn't hold; the mismatching actual and expected values are printed.

The downside of this change is the twofold increase in the number of
XFAIL messages. For example:
XFAIL  : TestCodeCompletion::testCompletionPriority(class) Derived class is not added to the Best Matches group
   Loc: [/path/to/test_codecompletion.cpp(250)]
XFAIL  : TestCodeCompletion::testCompletionPriority(class) 〃
   Loc: [/path/to/test_codecompletion.cpp(254)]