Remove const from childUrls return and annotate other consts
When a function returns a const value, this prevents moves and thus
might involve unnecessary copying. However, for functions whose return
values are used in for-range loops, the const prevents detachment, so is
desired. This commit removes the const from childUlrs (its not used for
iteration) and annotates the functions which should remain const.