Refactor ourProject as suggested by @igorkushnir
Instead of having a function to check whether a project is valid
(ourProject), we use a function which returns a pointer to the
QStandardItem corresponding to a valid project and returns nullptr for
invalid projects. This can be used in the same way, but in case the
project is valid, returns more information, which can then be reused.