ClangSupport: override indentationSample()
ILanguageSupport::indentationSample() is used in a single place:
CustomScriptPlugin::computeIndentationFromSample(), which in turn is
used only in CustomScriptPlugin::indentation(). When indentationSample()
returns an empty string (as the base class implementation does),
CustomScriptPlugin::indentation() fails and returns invalid Indentation.

f31d32f49458897c97388a4b8f40f97e5af02f59 removed CppLanguageSupport,
which was the only class that overrode indentationSample(). Not copying
this override into ClangSupport was probably a mistake. This commit
copies the last version of CppLanguageSupport::indentationSample() into
ClangSupport::indentationSample(). The copied sample matches the example
in the documentation for ILanguageSupport::indentationSample().