Add refresh action to git more commit toolview ctx menus
Before, when an element in one of the categories (staged/unstaged/...)
was selected, its context menu didn't have the refresh repo action. This
was not very convenient, since one had to first unselect the element and
then trigger the context menu again, when the repo model was out of date.
This commit fixes this minor annoyance.