CustomScriptPreferences: clarify ${PROJECT_NAME} description
Before analyzing the code, I thought that ${PROJECT_NAME} (verbatim) in
a formatter command is replaced by the path of the project, to which the
file being formatted belongs. In fact, the command must contain a
hardcoded name of some concrete project, which must be open for the
replacement to succeed. Obviously such a command cannot be usefully
shared between projects with different names. But at least the true
meaning of this very specialized variable is easier to understand now.