Enable the commit form after git commit failure
This fixes a bug which would lead to the commit form being disabled and
unusable when a commit failed for some reason.

The commit form was disabled when the commit job started and was meant to
be enabled when the job finished. However, the wrong signal was connected
to, which meant that in case of failure the code never ran.

BUG: 443475
FIXED-IN: 5.7.0