pick-session-shell: show the list of sessions in time
23ed228694be6e3082558d866a52e79c8923845a replaced endl with '\n' in many
places. Here the replacement caused a regression. The QTextStream must
be flushed immediately after the loop, before the user pick is read from
the standard input. Otherwise, the user has to pick a number at random
as the available session list is not displayed until after the number is
read from stdin.