CustomScriptPlugin: remove a temporary file after formatting
CustomScriptPlugin::formatSourceWithStyle() creates a code.* file in the
temporary directory if the current custom script style's command
contains $TMPFILE variable. One predefined style - kdev_format_source -
has this variable in its command. This temporary file is never used
after the function that creates it returns. Let us keep QTemporaryFile's
default auto-remove mode to save space in the temporary directory.

28770089a43024f520bd8e109f586512126ea555 added the line removed in this
commit. But I don't understand why it was beneficial back in 2011.
Perhaps it was a temporary change, which the commit's author forgot to
revert. Keeping the temporary file is most likely not useful today.