Add a new problem scope: DocumentsInPath
This allows us to set a specific path filter which can be used to
reduces the size of the problem set. Note that this is similar
to the existing proxy model filter, but has the big advantage
that we can filter earlier. Especially for large projects, this
allows us to filter once, instead of repeatedly whenever you are
editing a file.