Improve performance when queueing documents for parsing
I'm working with a project that has more than 1400 Unit-Tests and they are all in a single executable. This causes CTestFindJob to queue each c++ File compiled into the test executable more than 1400 times. We might probably want to change something about that, but that is a bit more work; Therefore I chose to get the low hanging fruits. These changes improve the queuing performance by a factor of approx. 10. For my project this means roughly 3 Seconds of waiting instead of more than 30 Seconds before the change.