Show an error message in the git commit view when committing fails
When committing fails, the user should be informed (and not expected to
notice this by seeing an error in the Version Control output view or knowing
commit message not being cleared means failure).

I opted to use a KMessageWidget instead of a Modal Dialog (per the HIG guidelines,
although they are a bit unclear, since there is no clear action that the
user can take to remedy the situation).

I also opted for just a short message pointing the user to the Version Control
output view for more information due to the limited space available in the
commit toolview.

BUG: 443475
FIXED-IN: 5.7.0